+++ app/ante/integration_test.go
@@ -1,7 +1,7 @@
 package ante_test
 
 import (
-	sdkmath "cosmossdk.io/math"
+	"cosmossdk.io/math"
 	cryptotypes "github.com/cosmos/cosmos-sdk/crypto/types"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	banktypes "github.com/cosmos/cosmos-sdk/x/bank/types"
@@ -57,9 +57,9 @@ var _ = Describe("when sending a Cosmos transaction", Label("AnteHandler"), Orde
 			rewards sdk.DecCoins
 			// minExpRewards are the minimun rewards that should be accrued
 			// for the test case
-			minExpRewards  = sdk.DecCoins{sdk.DecCoin{Amount: sdkmath.LegacyNewDec(1e5), Denom: s.network.GetDenom()}}
-			delegationCoin = sdk.Coin{Amount: sdkmath.NewInt(1e15), Denom: s.network.GetDenom()}
-			transferAmt    = sdkmath.NewInt(1e14)
+			minExpRewards  = sdk.DecCoins{sdk.DecCoin{Amount: math.LegacyNewDec(1e5), Denom: s.network.GetDenom()}}
+			delegationCoin = sdk.Coin{Amount: math.NewInt(1e15), Denom: s.network.GetDenom()}
+			transferAmt    = math.NewInt(1e14)
 		)
 
 		BeforeEach(func() {
@@ -85,14 +85,16 @@ var _ = Describe("when sending a Cosmos transaction", Label("AnteHandler"), Orde
 			prevBalanceRes, err := s.grpcHandler.GetBalance(addr, s.network.GetDenom())
 			Expect(err).To(BeNil())
 
-			baseFeeRes, err := s.grpcHandler.GetBaseFee()
+			baseFeeRes, err := s.grpcHandler.GetEvmBaseFee()
 			Expect(err).To(BeNil())
+
+			gasPrice := baseFeeRes.BaseFee
 
 			res, err := s.factory.ExecuteCosmosTx(
 				priv,
 				commonfactory.CosmosTxArgs{
 					Msgs:     []sdk.Msg{msg},
-					GasPrice: baseFeeRes.BaseFee,
+					GasPrice: gasPrice,
 				},
 			)
 			Expect(err).To(BeNil())
@@ -103,7 +105,7 @@ var _ = Describe("when sending a Cosmos transaction", Label("AnteHandler"), Orde
 			Expect(err).To(BeNil())
 
 			// fees should be deducted from balance
-			feesAmt := sdkmath.NewInt(res.GasWanted).Mul(*baseFeeRes.BaseFee)
+			feesAmt := math.NewInt(res.GasWanted).Mul(*gasPrice)
 			balanceRes, err := s.grpcHandler.GetBalance(addr, s.network.GetDenom())
 			Expect(err).To(BeNil())
 			Expect(balanceRes.Balance.Amount).To(Equal(prevBalanceRes.Balance.Amount.Sub(transferAmt).Sub(feesAmt)))
@@ -129,7 +131,7 @@ var _ = Describe("when sending a Cosmos transaction", Label("AnteHandler"), Orde
 				addr,
 				sdk.Coins{
 					sdk.Coin{
-						Amount: sdkmath.NewInt(1),
+						Amount: math.NewInt(1),
 						Denom:  s.network.GetDenom(),
 					},
 				},
@@ -141,7 +143,7 @@ var _ = Describe("when sending a Cosmos transaction", Label("AnteHandler"), Orde
 			msg = &banktypes.MsgSend{
 				FromAddress: addr.String(),
 				ToAddress:   "evmos1dx67l23hz9l0k9hcher8xz04uj7wf3yu26l2yn",
-				Amount:      sdk.Coins{sdk.Coin{Amount: sdkmath.NewInt(1e14), Denom: s.network.GetDenom()}},
+				Amount:      sdk.Coins{sdk.Coin{Amount: math.NewInt(1e14), Denom: s.network.GetDenom()}},
 			}
 		})
 
@@ -170,7 +172,7 @@ var _ = Describe("when sending a Cosmos transaction", Label("AnteHandler"), Orde
 	Context("and the sender account has not enough balance but sufficient staking rewards to pay for the transaction cost", func() {
 		// minExpRewards are the minimun rewards that should be accrued
 		// for the test case
-		minExpRewards := sdk.DecCoins{sdk.DecCoin{Amount: sdkmath.LegacyNewDec(1e8), Denom: s.network.GetDenom()}}
+		minExpRewards := sdk.DecCoins{sdk.DecCoin{Amount: math.LegacyNewDec(1e8), Denom: s.network.GetDenom()}}
 
 		BeforeEach(func() {
 			addr, priv = testutiltx.NewAccAddressAndKey()
@@ -182,7 +184,7 @@ var _ = Describe("when sending a Cosmos transaction", Label("AnteHandler"), Orde
 				addr,
 				sdk.Coins{
 					sdk.Coin{
-						Amount: sdkmath.NewInt(1e18),
+						Amount: math.NewInt(1e18),
 						Denom:  s.network.GetDenom(),
 					},
 				},
@@ -193,7 +195,7 @@ var _ = Describe("when sending a Cosmos transaction", Label("AnteHandler"), Orde
 
 			// delegate some tokens and make sure the remaining balance is not sufficient to cover the tx fees
 			valAddr := s.network.GetValidators()[1].OperatorAddress
-			err = s.factory.Delegate(priv, valAddr, sdk.NewCoin(s.network.GetDenom(), sdkmath.NewInt(9888e14)))
+			err = s.factory.Delegate(priv, valAddr, sdk.NewCoin(s.network.GetDenom(), math.NewInt(9888e14)))
 			Expect(err).To(BeNil())
 
 			_, err = integrationutils.WaitToAccrueRewards(s.network, s.grpcHandler, addr.String(), minExpRewards)
@@ -202,7 +204,7 @@ var _ = Describe("when sending a Cosmos transaction", Label("AnteHandler"), Orde
 			msg = &banktypes.MsgSend{
 				FromAddress: addr.String(),
 				ToAddress:   "evmos1dx67l23hz9l0k9hcher8xz04uj7wf3yu26l2yn",
-				Amount:      sdk.Coins{sdk.Coin{Amount: sdkmath.NewInt(1), Denom: s.network.GetDenom()}},
+				Amount:      sdk.Coins{sdk.Coin{Amount: math.NewInt(1), Denom: s.network.GetDenom()}},
 			}
 		})
 	})