+++ cmd/evmosd/testnet.go
@@ -75,7 +75,7 @@ type initArgs struct {
 	numValidators     int
 	outputDir         string
 	startingIPAddress string
-	baseFee           math.Int
+	baseFee           math.LegacyDec
 	minGasPrice       math.LegacyDec
 }
 
@@ -161,10 +161,12 @@ Example:
 			baseFee, _ := cmd.Flags().GetString(flagBaseFee)
 			minGasPrice, _ := cmd.Flags().GetString(flagMinGasPrice)
 
-			var ok bool
-			args.baseFee, ok = math.NewIntFromString(baseFee)
-			if !ok || args.baseFee.LT(math.ZeroInt()) {
-				return fmt.Errorf("invalid value for --base-fee. expected a int number greater than or equal to 0 but got %s", baseFee)
+			args.baseFee, err = math.LegacyNewDecFromStr(baseFee)
+			if err != nil {
+				return fmt.Errorf("invalid value for --base-fee. failed during the conversion from string %s to decimal. err %s", baseFee, err.Error())
+			}
+			if args.baseFee.IsNil() || args.baseFee.IsNegative() {
+				return fmt.Errorf("invalid value for --base-fee. expected a int or decimal number greater than or equal to 0 but got %s", baseFee)
 			}
 
 			args.minGasPrice, err = math.LegacyNewDecFromStr(minGasPrice)
@@ -348,8 +350,8 @@ func initTestnetFiles(
 		}
 
 		minGasPrice := args.minGasPrice
-		if math.LegacyNewDecFromInt(args.baseFee).GT(args.minGasPrice) {
-			minGasPrice = math.LegacyNewDecFromInt(args.baseFee)
+		if args.baseFee.GT(args.minGasPrice) {
+			minGasPrice = args.baseFee
 		}
 
 		txBuilder.SetMemo(memo)
@@ -412,7 +414,7 @@ func initGenFiles(
 	genBalances []banktypes.Balance,
 	genFiles []string,
 	numValidators int,
-	baseFee math.Int,
+	baseFee math.LegacyDec,
 	minGasPrice math.LegacyDec,
 ) error {
 	appGenState := mbm.DefaultGenesis(clientCtx.Codec)