+++ tests/nix_tests/cosmoscli.py
@@ -13,6 +13,7 @@ DEFAULT_GAS = "250000"
 
 
 class ChainCommand:
+
     def __init__(self, cmd):
         self.cmd = cmd
 
@@ -1154,7 +1155,7 @@ class CosmosCLI:
         if not base_fee:
             raise ValueError(f"failed to return base fee: {out_dict}")
 
-        return int(base_fee)
+        return float(base_fee)
 
     # ==========================
     #        AUTHZ Module