+++ testutil/tx/eth.go
@@ -103,6 +103,12 @@ func CreateEthTx(
 	fromAddr := common.BytesToAddress(from.Bytes())
 	chainID := evmtypes.GetChainConfig().ChainID
 
+	baseFeeRes, err := appEvmos.EvmKeeper.BaseFee(ctx, &evmtypes.QueryBaseFeeRequest{})
+	if err != nil {
+		return nil, errorsmod.Wrap(err, "failed to get fee during eth tx creation")
+	}
+	baseFee := baseFeeRes.BaseFee.BigInt()
+
 	// When we send multiple Ethereum Tx's in one Cosmos Tx, we need to increment the nonce for each one.
 	nonce := appEvmos.EvmKeeper.GetNonce(ctx, fromAddr) + uint64(nonceIncrement) //nolint:gosec // G115
 	evmTxParams := &evmtypes.EvmTxArgs{
@@ -111,7 +117,7 @@ func CreateEthTx(
 		To:        &toAddr,
 		Amount:    amount,
 		GasLimit:  100000,
-		GasFeeCap: appEvmos.FeeMarketKeeper.GetBaseFee(ctx),
+		GasFeeCap: baseFee,
 		GasTipCap: big.NewInt(1),
 		Accesses:  &ethtypes.AccessList{},
 	}