+++ x/evm/keeper/abci.go
@@ -5,9 +5,34 @@ package keeper
 import (
 	storetypes "cosmossdk.io/store/types"
 	sdk "github.com/cosmos/cosmos-sdk/types"
+	"github.com/evmos/evmos/v20/x/evm/types"
 
 	ethtypes "github.com/ethereum/go-ethereum/core/types"
 )
+
+// BeginBlock emits a base fee event which will be adjusted to the evm decimals
+func (k *Keeper) BeginBlock(ctx sdk.Context) error {
+	logger := ctx.Logger().With("begin_block", "evm")
+
+	// Base fee is already set on FeeMarket BeginBlock
+	// that runs before this one
+	// We emit this event on the EVM and FeeMarket modules
+	// because they can be different if the evm denom has 6 decimals
+	res, err := k.BaseFee(ctx, &types.QueryBaseFeeRequest{})
+	if err != nil {
+		logger.Error("error when getting base fee", "error", err.Error())
+	}
+	if res != nil && res.BaseFee != nil && !res.BaseFee.IsNil() {
+		// Store current base fee in event
+		ctx.EventManager().EmitEvents(sdk.Events{
+			sdk.NewEvent(
+				types.EventTypeFeeMarket,
+				sdk.NewAttribute(types.AttributeKeyBaseFee, res.BaseFee.String()),
+			),
+		})
+	}
+	return nil
+}
 
 // EndBlock also retrieves the bloom filter value from the transient store and commits it to the
 // KVStore. The EVM end block logic doesn't update the validator set, thus it returns