+++ x/evm/types/denom_config.go
@@ -17,16 +17,32 @@ import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
 )
 
-// Decimals is a wrapper around uint32 to represent the decimal representation
-// of a Cosmos coin.
-type Decimals uint32
-
+// NOTE: Remember to add the ConversionFactor associated with constants.
 const (
 	// SixDecimals is the Decimals used for Cosmos coin with 6 decimals.
 	SixDecimals Decimals = 6
 	// EighteenDecimals is the Decimals used for Cosmos coin with 18 decimals.
 	EighteenDecimals Decimals = 18
 )
+
+// Decimals is a wrapper around uint32 to represent the decimal representation
+// of a Cosmos coin.
+type Decimals uint32
+
+// ConversionFactor returns the conversion factor between the Decimals value and
+// the 18 decimals representation, i.e. `EighteenDecimals`.
+//
+// NOTE: This function does not check if the Decimal instance is valid or
+// not and by default returns the conversion factor of 1, i.e. from 18 decimals
+// to 18 decimals. We cannot have a non supported Decimal since it is checked
+// and validated.
+func (d Decimals) ConversionFactor() math.Int {
+	if d == SixDecimals {
+		return math.NewInt(1e12)
+	}
+
+	return math.NewInt(1)
+}
 
 // EvmCoinInfo struct holds the name and decimals of the EVM denom. The EVM denom
 // is the token used to pay fees in the EVM.
@@ -76,18 +92,4 @@ func setEVMCoinInfo(evmdenom EvmCoinInfo) {
 	evmCoinInfo = new(EvmCoinInfo)
 	setEVMCoinDenom(evmdenom.Denom)
 	setEVMCoinDecimals(evmdenom.Decimals)
-}
-
-// ConversionFactor returns the conversion factor between the Decimals value and
-// the 18 decimals representation, i.e. `EighteenDecimals`.
-//
-// NOTE: This function does not check if the Decimal instance is valid or
-// not and by default returns the conversion factor of 1, i.e. from 18 decimals
-// to 18 decimals.
-func (d Decimals) ConversionFactor() math.Int {
-	if d == SixDecimals {
-		return math.NewInt(1e12)
-	}
-
-	return math.NewInt(1)
 }