+++ x/feemarket/keeper/eip1559.go
@@ -3,11 +3,9 @@
 package keeper
 
 import (
-	"math/big"
-
+	sdkmath "cosmossdk.io/math"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 
-	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/common/math"
 )
 
@@ -15,12 +13,12 @@ import (
 // block during BeginBlock. If the NoBaseFee parameter is enabled or below activation height, this function returns nil.
 // NOTE: This code is inspired from the go-ethereum EIP1559 implementation and adapted to Cosmos SDK-based
 // chains. For the canonical code refer to: https://github.com/ethereum/go-ethereum/blob/master/consensus/misc/eip1559.go
-func (k Keeper) CalculateBaseFee(ctx sdk.Context) *big.Int {
+func (k Keeper) CalculateBaseFee(ctx sdk.Context) sdkmath.LegacyDec {
 	params := k.GetParams(ctx)
 
 	// Ignore the calculation if not enabled
 	if !params.IsBaseFeeEnabled(ctx.BlockHeight()) {
-		return nil
+		return sdkmath.LegacyDec{}
 	}
 
 	consParams := ctx.ConsensusParams()
@@ -29,66 +27,69 @@ func (k Keeper) CalculateBaseFee(ctx sdk.Context) *big.Int {
 	// defined in the parameters (DefaultBaseFee if it hasn't been changed by
 	// governance).
 	if ctx.BlockHeight() == params.EnableHeight {
-		return params.BaseFee.BigInt()
+		return params.BaseFee
 	}
 
 	// get the block gas used and the base fee values for the parent block.
 	// NOTE: this is not the parent's base fee but the current block's base fee,
 	// as it is retrieved from the transient store, which is committed to the
 	// persistent KVStore after EndBlock (ABCI Commit).
-	parentBaseFee := params.BaseFee.BigInt()
-	if parentBaseFee == nil {
-		return nil
+	parentBaseFee := params.BaseFee
+	if parentBaseFee.IsNil() {
+		return sdkmath.LegacyDec{}
 	}
 
 	parentGasUsed := k.GetBlockGasWanted(ctx)
 
-	gasLimit := new(big.Int).SetUint64(math.MaxUint64)
+	gasLimit := sdkmath.NewIntFromUint64(math.MaxUint64)
 
 	// NOTE: a MaxGas equal to -1 means that block gas is unlimited
 	if consParams.Block != nil && consParams.Block.MaxGas > -1 {
-		gasLimit = big.NewInt(consParams.Block.MaxGas)
+		gasLimit = sdkmath.NewInt(consParams.Block.MaxGas)
 	}
 
 	// CONTRACT: ElasticityMultiplier cannot be 0 as it's checked in the params
 	// validation
-	parentGasTargetBig := new(big.Int).Div(gasLimit, new(big.Int).SetUint64(uint64(params.ElasticityMultiplier)))
-	if !parentGasTargetBig.IsUint64() {
-		return nil
+	parentGasTargetInt := gasLimit.Quo(sdkmath.NewIntFromUint64(uint64(params.ElasticityMultiplier)))
+	if !parentGasTargetInt.IsUint64() {
+		return sdkmath.LegacyDec{}
 	}
 
-	parentGasTarget := parentGasTargetBig.Uint64()
-	baseFeeChangeDenominator := new(big.Int).SetUint64(uint64(params.BaseFeeChangeDenominator))
+	parentGasTarget := parentGasTargetInt.Uint64()
+	baseFeeChangeDenominator := sdkmath.NewIntFromUint64(uint64(params.BaseFeeChangeDenominator))
 
 	// If the parent gasUsed is the same as the target, the baseFee remains
 	// unchanged.
 	if parentGasUsed == parentGasTarget {
-		return new(big.Int).Set(parentBaseFee)
+		return parentBaseFee
+	}
+
+	if parentGasTargetInt.IsZero() {
+		return sdkmath.LegacyZeroDec()
 	}
 
 	if parentGasUsed > parentGasTarget {
 		// If the parent block used more gas than its target, the baseFee should
 		// increase.
-		gasUsedDelta := new(big.Int).SetUint64(parentGasUsed - parentGasTarget)
-		x := new(big.Int).Mul(parentBaseFee, gasUsedDelta)
-		y := x.Div(x, parentGasTargetBig)
-		baseFeeDelta := math.BigMax(
-			x.Div(y, baseFeeChangeDenominator),
-			common.Big1,
+		gasUsedDelta := sdkmath.NewIntFromUint64(parentGasUsed - parentGasTarget)
+		x := parentBaseFee.MulInt(gasUsedDelta)
+		y := x.QuoInt(parentGasTargetInt)
+		baseFeeDelta := sdkmath.LegacyMaxDec(
+			y.QuoInt(baseFeeChangeDenominator),
+			sdkmath.LegacyOneDec(),
 		)
 
-		return x.Add(parentBaseFee, baseFeeDelta)
+		return parentBaseFee.Add(baseFeeDelta)
 	}
 
 	// Otherwise if the parent block used less gas than its target, the baseFee
 	// should decrease.
-	gasUsedDelta := new(big.Int).SetUint64(parentGasTarget - parentGasUsed)
-	x := new(big.Int).Mul(parentBaseFee, gasUsedDelta)
-	y := x.Div(x, parentGasTargetBig)
-	baseFeeDelta := x.Div(y, baseFeeChangeDenominator)
+	gasUsedDelta := sdkmath.NewIntFromUint64(parentGasTarget - parentGasUsed)
+	x := parentBaseFee.MulInt(gasUsedDelta)
+	y := x.QuoInt(parentGasTargetInt)
+	baseFeeDelta := y.QuoInt(baseFeeChangeDenominator)
 
 	// Set global min gas price as lower bound of the base fee, transactions below
 	// the min gas price don't even reach the mempool.
-	minGasPrice := params.MinGasPrice.TruncateInt().BigInt()
-	return math.BigMax(x.Sub(parentBaseFee, baseFeeDelta), minGasPrice)
+	return sdkmath.LegacyMaxDec(parentBaseFee.Sub(baseFeeDelta), params.MinGasPrice)
 }