+++ x/feemarket/keeper/grpc_query.go
@@ -30,11 +30,7 @@ func (k Keeper) BaseFee(c context.Context, _ *types.QueryBaseFeeRequest) (*types
 
 	res := &types.QueryBaseFeeResponse{}
 	baseFee := k.GetBaseFee(ctx)
-
-	if baseFee != nil {
-		aux := sdkmath.NewIntFromBigInt(baseFee)
-		res.BaseFee = &aux
-	}
+	res.BaseFee = &baseFee
 
 	return res, nil
 }